requirements.txt
pyserial==3.5
python-socketio==5.11.2
numpy==1.26.4
matplotlib==3.8.4
PyQt5==5.15.10